# public/openapi.yaml
openapi: 3.1.0
info:
  title: Mira Studio API
  description: Gateway for the Mira experience engine and Goal OS. GPT operations go through 7 endpoints. All /create and /update payloads are FLAT (no nesting under a "payload" key).
  version: 2.2.0
servers:
  - url: https://mira-maddyup.vercel.app/  
    description: Mira Studio Backend
paths:
...
  /api/gpt/create:
    post:
      operationId: createEntity
      summary: Create experiences, ideas, goals, steps, knowledge, skill domains, map objects, memories, or boards
      description: |
        FLAT payload — all fields alongside `type`. Do NOT nest under a `payload` key.
        Call GET /api/gpt/discover?capability=<type> for the exact schema.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [experience, ephemeral, idea, step, goal, knowledge, skill_domain, map_node, map_edge, map_cluster, memory, board, board_from_text]
                  description: The entity type to create.
                templateId:
                  type: string
                  description: UUID of the experience template. Call discover?capability=templates for valid IDs.
                userId:
                  type: string
                  default: "a0000000-0000-0000-0000-000000000001"
                title:
                  type: string
                  description: Title of the experience, idea, or step.
                goal:
                  type: string
                  description: For experiences — what the user will achieve.
                description:
                  type: string
                  description: For goals/memories/nodes — content or summary.
                resolution:
                  type: object
                  properties:
                    depth:
                      type: string
                      enum: [light, medium, heavy]
                    mode:
                      type: string
                      enum: [illuminate, practice, challenge, build, reflect, study]
                    timeScope:
                      type: string
                      enum: [immediate, session, multi_day, ongoing]
                    intensity:
                      type: string
                      enum: [low, medium, high]
                reentry:
                  type: object
                  properties:
                    trigger:
                      type: string
                      enum: [time, completion, inactivity, manual]
                    prompt:
                      type: string
                    contextScope:
                      type: string
                      enum: [minimal, full, focused]
                steps:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [lesson, challenge, reflection, questionnaire, essay_tasks, plan_builder, checkpoint]
                      title:
                        type: string
                      payload:
                        type: object
                        additionalProperties: true
                      blocks:
                        type: array
                        items:
                          type: object
                          additionalProperties: true
                  description: Array of steps for the experience.
                curriculum_outline_id:
                  type: string
                previousExperienceId:
                  type: string
                domains:
                  type: array
                  items:
                    type: string
                goalId:
                  type: string
                name:
                  type: string
                  description: REQUIRED for skill_domain or board.
                purpose:
                  type: string
                  enum: [general, idea_planning, curriculum_review, lesson_plan, research_tracking, strategy]
                  description: For type=board.
                linkedEntityId:
                  type: string
                linkedEntityType:
                  type: string
                  enum: [goal, experience, knowledge]
                kind:
                  type: string
                  enum: [observation, strategy, idea, preference, tactic, assessment, note]
                  description: For type=memory.
                memory_class:
                  type: string
                  enum: [semantic, episodic, procedural]
                  description: For type=memory.
                topic:
                  type: string
                  description: REQUIRED for memory or outline.
                content:
                  type: string
                  description: For type=memory/knowledge/map_node.
                tags:
                  type: array
                  items:
                    type: string
                rawPrompt:
                  type: string
                gptSummary:
                  type: string
                prompt:
                  type: string
                  description: For type=board_from_text.
                experienceId:
                  type: string
                boardId:
                  type: string
                label:
                  type: string
                color:
                  type: string
                position_x:
                  type: number
                position_y:
                  type: number
                sourceNodeId:
                  type: string
                targetNodeId:
                  type: string
                centerNode:
                  type: object
                childNodes:
                  type: array
                  items:
                    type: object
                step_type:
                  type: string
                  enum: [lesson, challenge, reflection, questionnaire, essay_tasks, plan_builder, checkpoint]
                sections:
                  type: array
                  items:
                    type: object
                prompts:
                  type: array
                  items:
                    type: object
                questions:
                  type: array
                  items:
                    type: object
                tasks:
                  type: array
                  items:
                    type: object
                knowledge_unit_id:
                  type: string
                passing_threshold:
                  type: integer
                on_fail:
                  type: string
                payload:
                  type: object
                blocks:
                  type: array
                  items:
                    type: object
      responses:
        '201':
          description: Created
        '400':
          description: Validation error
...
  /api/gpt/memory:
    get:
      operationId: listMemories
      summary: Query persistent agent memories
      description: Query the "Notebook" layer of agent memory. Entries are filtered by kind, topic, or content.
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: Optional User ID.
        - name: kind
          in: query
          required: false
          schema:
            type: string
          description: Filter by memory kind (observation, preference, etc).
        - name: topic
          in: query
          required: false
          schema:
            type: string
          description: Filter by topic string.
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Keyword search in memory content.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                  totalCount:
                    type: integer
                  lastRecordedAt:
                    type: string
                    nullable: true
...